re-declare obsolete method UpnpSetContentLength, for binary compatibility with previous libupnp version. + correct type of g_maxContentLength to size_t
git-svn-id: https://svn.code.sf.net/p/djmount/code/trunk@114 d2b06c04-d219-0410-9cba-8d529fbc48d6

+++ libupnp/upnp/src/inc/upnpapi.h
@@ -54,7 +54,7 @@
 
 #define DEFAULT_MAXAGE 1800
 
-extern int g_maxContentLength;
+extern size_t g_maxContentLength;
 
 // 30-second timeout
 #define UPNP_TIMEOUT (30 + MINIMUM_DELAY)
